requirements.txt
git+https://github.com/NVIDIA/apex
munch
adamp
fire
pyyaml